Add distributed training support for ensembles
When creating ensembles using create_ensemble, every member config
that is not already trained is written to a request pool on disk.
The process that called create_ensemble will then start to fetch
members to train from this pool, but other processes can also help out.

lib.distributed_trainer is a simple looping process that picks up work when
available.